changed string bind length to uint32_t from uint16_t 	added clearBinds() call after executeQuery() in perl dbi driver

+++ src/connection/sqlrcursor/binds.C
@@ -108,7 +108,7 @@ bool sqlrcursor_svr::handleBinds() {
 bool sqlrcursor_svr::inputBindString(const char *variable,
 					uint16_t variablesize,
 					const char *value,
-					uint16_t valuesize,
+					uint32_t valuesize,
 					int16_t *isnull) {
 	// by default, do nothing...
 	return true;
