+++ src/connections/firebird/firebirdconnection.C
@@ -282,7 +282,7 @@ bool firebirdcursor::prepareQuery(const char *query, uint32_t length) {
 bool firebirdcursor::inputBindString(const char *variable,
 					uint16_t variablesize,
 					const char *value,
-					uint16_t valuesize,
+					uint32_t valuesize,
 					int16_t *isnull) {
 
 	// make bind vars 1 based like all other db's
