+++ src/connections/postgresql/postgresqlconnection.h
@@ -33,7 +33,7 @@ class postgresqlcursor : public sqlrcursor_svr {
 		bool		inputBindString(const char *variable, 
 						uint16_t variablesize,
 						const char *value, 
-						uint16_t valuesize,
+						uint32_t valuesize,
 						int16_t *isnull);
 		bool		inputBindInteger(const char *variable, 
 						uint16_t variablesize,
